debugging the connection of splitter to the compilationsuit
